+++ src/fields/SoSFInt32.cpp
@@ -73,7 +73,7 @@ SoSFInt32::readValue(SoInput * in)
 void
 sosfint32_write_value(SoOutput * out, int32_t val)
 {
-  out->write(val);
+  out->write((int) val);
 }
 
 void
